yeaaah onemax and engine is pretty much done i thinik

+++ src/main/java/subsym/onemax/OneMaxGenotype.java
@@ -18,7 +18,7 @@ public class OneMaxGenotype extends Genotype implements Phenotype {
   }
 
   @Override
-  public int fitness() {
+  public double fitness() {
     return getBits().cardinality();
   }
